+++ src/data/consume/choker-arm-mutagen.ts
@@ -1,0 +1,106 @@
+import { AssistantAction } from "action.ts";
+import { EffectSource } from "foundry-pf2e";
+import { AssistantMessage } from "message.ts";
+
+export const actions: AssistantAction[] = [
+    {
+        trigger: "consume",
+        predicate: ["consumable:choker-arm-mutagen-lesser"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.FDSl6DFblUjITOgP", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+    {
+        trigger: "consume",
+        predicate: ["consumable:choker-arm-mutagen-moderate"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.IBu1RWKW5JX70nse", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+    {
+        trigger: "consume",
+        predicate: ["consumable:choker-arm-mutagen-greater"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-effects.Item.HlUL7FmGLyS35Jpp", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+    {
+        trigger: "consume",
+        predicate: ["consumable:choker-arm-mutagen-major"],
+        process: async (message: AssistantMessage) => {
+            if (!message.speaker?.actor) return;
+            const target = message.target?.actor ?? message.speaker.actor;
+
+            game.assistant.socket.addEmbeddedItem(target, "Compendium.pf2e.equipment-srd.Item.bvIVxDq1wh6IavHP", {
+                _id: null,
+                system: {
+                    context: {
+                        origin: {
+                            actor: message.speaker.actor.uuid,
+                            token: message.speaker.token?.uuid ?? null,
+                            item: message.item?.uuid ?? null,
+                            spellcasting: null,
+                            rollOptions: message.item?.getOriginData().rollOptions ?? [],
+                        },
+                        target: null,
+                        roll: null,
+                    },
+                },
+            } as EffectSource);
+        },
+    },
+];